+++ RELEASENOTES.md
@@ -30,6 +30,9 @@
     *   Enable sending `CmcdData` for manifest requests in adaptive streaming
         formats DASH, HLS, and SmoothStreaming
         ([#1951](https://github.com/androidx/media/issues/1951)).
+    *   Provide `MediaCodecInfo` of the codec that will be initialized in
+        `MediaCodecRenderer.onReadyToInitializeCodec`
+        ([#1963](https://github.com/androidx/media/pull/1963)).
 *   Transformer:
     *   Update parameters of `VideoFrameProcessor.registerInputStream` and
         `VideoFrameProcessor.Listener.onInputStreamRegistered` to use `Format`.